feat: implementacao do endpoint de perfil de risco por clienteId

+++ src/main/java/br/gov/caixa/domain/constants/ApiConstants.java
@@ -1,10 +1,12 @@
 package br.gov.caixa.domain.constants;
 
+import java.math.BigDecimal;
+
 public class ApiConstants {
 
-    public final static Double PONTUACAO_PERFIL_CONSERVADOR = 25.0;
-    public final static Double PONTUACAO_PERFIL_MODERADO = 50.0;
-    public final static Double PONTUACAO_PERFIL_AGRESSIVO = 75.0;
+    public final static BigDecimal PONTUACAO_PERFIL_CONSERVADOR = BigDecimal.valueOf(25.0);
+    public final static BigDecimal PONTUACAO_PERFIL_MODERADO = BigDecimal.valueOf(50.0);
+    public final static BigDecimal PONTUACAO_PERFIL_AGRESSIVO = BigDecimal.valueOf(75.0);
     public final static String TITULO_PERFIL_CONSERVADOR= "Conservador";
     public final static String TITULO_PERFIL_MODERADO = "Moderado";
     public final static String TITULO_PERFIL_AGRESSIVO = "Agressivo";
