+++ src/main/java/br/gov/caixa/domain/entity/AbstractEntity.java
@@ -1,13 +1,11 @@
 package br.gov.caixa.domain.entity;
 
-import io.quarkus.hibernate.orm.panache.PanacheEntity;
 import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 import jakarta.persistence.*;
 import lombok.Data;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import java.math.BigInteger;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -20,7 +18,7 @@ public class AbstractEntity extends PanacheEntityBase {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(columnDefinition = "BIGINT")
-    private BigInteger id;
+    private Long id;
 
     @Column(name = "uuid", unique = true, nullable = false, updatable = false)
     private String uuid;
